streamline blocks inside country detail
